requirements.txt
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
idna==2.8
lxml==4.3.3
oauthlib==3.0.1
pbr==5.2.0
prettytable==0.7.2
progress==1.5
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
Tempita==0.5.2
uritemplate==3.0.0
urllib3==1.24.3
